Remove Sentinel class
A sentinel monitoring tool may require to execute regular Redis commands that are supported by Sentinel nodes.
Including SentinelCommands as before seems to be the easy solution for supporting the scenario.
