+++ docs/3to4.md
@@ -128,8 +128,6 @@
 - Jedis.pubsubNumSub(String...) method now returns `Map<String, Long>` instead of
   `Map<String, String>`.
 
-- Sentinel methods have been removed from Jedis class and can be found in the new Sentinel class.
-
 - `setDataSource` method in Jedis class now has `protected` access.
 
 - JedisPoolAbstract class has been removed. Use Pool<Jedis>.
@@ -158,7 +156,7 @@
   - EVAL_RESULT (use ENCODED_OBJECT)
   - EVAL_BINARY_RESULT (use RAW_OBJECT)
 
-- All String variables representing Sentinel, Cluster and PubSub subcommands in Protocol class
+- All String variables representing Cluster, Sentinel and PubSub subcommands in Protocol class
   have been removed.
 
 - ClientKillParams.Type has been removed. Use ClientType.
