+++ src/test/java/redis/clients/jedis/util/JedisSentinelTestUtil.java
@@ -4,13 +4,13 @@ import java.util.concurrent.atomic.AtomicReference;
 
 import redis.clients.jedis.HostAndPort;
 import redis.clients.jedis.JedisPubSub;
-import redis.clients.jedis.Sentinel;
+import redis.clients.jedis.Jedis;
 import redis.clients.jedis.exceptions.FailoverAbortedException;
 
 public class JedisSentinelTestUtil {
 
   public static HostAndPort waitForNewPromotedMaster(final String masterName,
-      final Sentinel sentinelJedis, final Sentinel commandJedis) throws InterruptedException {
+      final Jedis sentinelJedis, final Jedis commandJedis) throws InterruptedException {
 
     final AtomicReference<String> newmaster = new AtomicReference<String>("");
